Extract teams-related models to their own module
