+++ src/api/teams/list.rs
@@ -33,7 +33,7 @@ impl<'octo, 'r> ListTeamsBuilder<'octo, 'r> {
     }
 
     /// Sends the actual request.
-    pub async fn send(self) -> Result<Page<models::RequestedTeam>> {
+    pub async fn send(self) -> Result<Page<models::teams::RequestedTeam>> {
         let url = format!("/orgs/{owner}/teams", owner = self.handler.owner);
         self.handler.crab.get(url, Some(&self)).await
     }
